renamed from strip to complement as that is what it does.
