+++ lib/index.js
@@ -2,7 +2,7 @@ module.exports = {
     iterateKeySet: require('./iterateKeySet'),
     toTree: require('./toTree'),
     toTreeWithUnion: require('./toTreeWithUnion'),
-    stripFromPathsWithTree: require('./stripFromPathsWithTree'),
+    pathsComplementFromTree: require('./pathsComplementFromTree'),
     toPaths: require('./toPaths'),
     collapse: require('./collapse')
 };
