+++ lib/pathsComplementFromTree.js
